Extract the atom data library to its own package
