+++ docs/manifest.json
@@ -1794,6 +1794,12 @@
 		"parent": "packages"
 	},
 	{
+		"title": "@wordpress/stan",
+		"slug": "packages-stan",
+		"markdown_source": "../packages/stan/README.md",
+		"parent": "packages"
+	},
+	{
 		"title": "@wordpress/token-list",
 		"slug": "packages-token-list",
 		"markdown_source": "../packages/token-list/README.md",
