+++ packages/data/README.md
@@ -345,14 +345,6 @@ _Returns_
 
 Undocumented declaration.
 
-<a name="createAtom" href="#createAtom">#</a> **createAtom**
-
-Undocumented declaration.
-
-<a name="createAtomRegistry" href="#createAtomRegistry">#</a> **createAtomRegistry**
-
-Undocumented declaration.
-
 <a name="createReduxStore" href="#createReduxStore">#</a> **createReduxStore**
 
 Creates a namespace object with a store derived from the reducer given.
@@ -365,10 +357,6 @@ _Parameters_
 _Returns_
 
 -   (unknown type): Store Object.
-
-<a name="createDerivedAtom" href="#createDerivedAtom">#</a> **createDerivedAtom**
-
-Undocumented declaration.
 
 <a name="createRegistry" href="#createRegistry">#</a> **createRegistry**
 
@@ -458,10 +446,6 @@ _Parameters_
 _Returns_
 
 -   `Function`: Registry selector that can be registered with a store.
-
-<a name="createStoreAtom" href="#createStoreAtom">#</a> **createStoreAtom**
-
-Undocumented declaration.
 
 <a name="dispatch" href="#dispatch">#</a> **dispatch**
