+++ packages/data/package.json
@@ -31,6 +31,7 @@
 		"@wordpress/is-shallow-equal": "file:../is-shallow-equal",
 		"@wordpress/priority-queue": "file:../priority-queue",
 		"@wordpress/redux-routine": "file:../redux-routine",
+		"@wordpress/stan": "file:../stan",
 		"equivalent-key-map": "^0.2.2",
 		"is-promise": "^4.0.0",
 		"lodash": "^4.17.19",
