+++ packages/keyboard-shortcuts/package.json
@@ -27,6 +27,7 @@
 		"@wordpress/data": "file:../data",
 		"@wordpress/element": "file:../element",
 		"@wordpress/keycodes": "file:../keycodes",
+		"@wordpress/stan": "file:../stan",
 		"lodash": "^4.17.19",
 		"rememo": "^3.0.0"
 	},
