+++ packages/stan/README.md
@@ -1,0 +1,78 @@
+# Stan
+
+A library to manage state.
+
+## Installation
+
+Install the module
+
+```bash
+npm install @wordpress/stan --save
+```
+
+_This package assumes that your code will run in an **ES2015+** environment. If you're using an environment that has limited or no support for ES2015+ such as lower versions of IE then using [core-js](https://github.com/zloirock/core-js) or [@babel/polyfill](https://babeljs.io/docs/en/next/babel-polyfill) will add support for these methods. Learn more about it in [Babel docs](https://babeljs.io/docs/en/next/caveats)._
+
+## Usage
+
+<!-- START TOKEN(Autogenerated API docs) -->
+
+<a name="createAtom" href="#createAtom">#</a> **createAtom**
+
+Creates a basic atom.
+
+_Parameters_
+
+-   _initialValue_ `T`: Initial Value in the atom.
+-   _id_ `[string]`: Atom id.
+
+_Returns_
+
+-   (unknown type): Createtd atom.
+
+<a name="createAtomRegistry" href="#createAtomRegistry">#</a> **createAtomRegistry**
+
+Creates a new Atom Registry.
+
+_Parameters_
+
+-   _onAdd_ `RegistryListener`: 
+-   _onDelete_ `RegistryListener`: 
+
+_Returns_
+
+-   (unknown type): Atom Registry.
+
+<a name="createDerivedAtom" href="#createDerivedAtom">#</a> **createDerivedAtom**
+
+Creates a derived atom.
+
+_Parameters_
+
+-   _resolver_ (unknown type): Atom Resolver.
+-   _updater_ (unknown type): Atom updater.
+-   _isAsync_ `[boolean]`: Atom resolution strategy.
+-   _id_ `[string]`: Atom id.
+
+_Returns_
+
+-   (unknown type): Createtd atom.
+
+<a name="createStoreAtom" href="#createStoreAtom">#</a> **createStoreAtom**
+
+Creates a store atom.
+
+_Parameters_
+
+-   _get_ (unknown type): Get the state value.
+-   _subscribe_ (unknown type): Subscribe to state changes.
+-   _dispatch_ (unknown type): Dispatch store changes,
+-   _id_ `[string]`: Atom id.
+
+_Returns_
+
+-   (unknown type): Store Atom.
+
+
+<!-- END TOKEN(Autogenerated API docs) -->
+
+<br/><br/><p align="center"><img src="https://s.w.org/style/images/codeispoetry.png?1" alt="Code is Poetry." /></p>
